requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-csv
celery
django-celery-beat
redis
